+++ cts/include/huaweicloud/cts/v3/CtsRegion.h
@@ -168,6 +168,10 @@ public:
             { 
                 "https://cts.ru-moscow-1.myhuaweicloud.com"
             });
+        regions["ae-ad-1"] = Region("ae-ad-1",
+            { 
+                "https://cts.ae-ad-1.myhuaweicloud.com"
+            });
         return regions;
     }
 private: