+++ kvs/include/huaweicloud/kvs/v1/model/ScanKvRequestBody.h
@@ -97,6 +97,24 @@ public:
     void unsetfilterExpression();
     void setFilterExpression(const Condition_expression& value);
 
+    /// <summary>
+    /// 对表进行采样，尽最大努力保证返回的段列表均分整张表。举例：sample_segments_count&#x3D;4，返回的段列表[MinKey, KV1)、[KV1,KV2)、[KV2,KV3)和[KV3,MaxKey)。用户可以使用scan-kv对这四个分区执行并发扫描，提高遍历效率。 - 范围: [1, 10000]。默认值：不执行采样。 - sample_segments_count仅能和table_name、start_key和end_key字段配合使用。Range分区模式下支持全表采样和范围采样；Hash分区模式仅支持全表扫描。 - 仅支持对Primary key进行采样，不支持本地/全局二级索引。 - 返回的段列表仅包含主键，不包含键值；且段列表是编码后的数据 ，不可修改。
+    /// </summary>
+
+    int32_t getSampleSegmentsCount() const;
+    bool sampleSegmentsCountIsSet() const;
+    void unsetsampleSegmentsCount();
+    void setSampleSegmentsCount(int32_t value);
+
+    /// <summary>
+    /// 返回查询条件对应的KV总数. - 当KV总数小于limit条件时，返回KV查询结果和KV总数。 - 当KV总数多于limit条件时，只返回KV总数。
+    /// </summary>
+
+    bool isReturnCountOnly() const;
+    bool returnCountOnlyIsSet() const;
+    void unsetreturnCountOnly();
+    void setReturnCountOnly(bool value);
+
 
 protected:
     std::string tableName_;
@@ -111,6 +129,10 @@ protected:
     bool endKeyIsSet_;
     Condition_expression filterExpression_;
     bool filterExpressionIsSet_;
+    int32_t sampleSegmentsCount_;
+    bool sampleSegmentsCountIsSet_;
+    bool returnCountOnly_;
+    bool returnCountOnlyIsSet_;
 
 };